services/ai-voice-connector/requirements.txt
# AI Voice Connector Service Dependencies

# Core framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
grpcio==1.60.0
grpcio-tools==1.60.0
pydantic==2.5.0

# OpenSIPS integration
opensips==0.1.5
sipmessage==0.5.0

# Audio processing
pyaudio==0.2.13
soundfile==0.12.1
numpy==1.24.3

# Async utilities
aiofiles==23.2.1
httpx==0.25.2

# Configuration
python-dotenv==1.0.0
pydantic-settings==2.1.0

# Monitoring
prometheus-client==0.19.0
structlog==23.2.0